requirements.txt
daphne==4.1.2
Django==5.0.7
mysqlclient==2.2.4
requests==2.32.3
python-decouple==3.8